td_api/requirements.txt
fastapi
scikit-learn
deta
python-dotenv